Add a helper script to find rustc archives
The script looks for rustc archives on the server. Archives are sorted
by dates, so the script iterates over days in a month to find them.

When it finds an archive, it downloads it in /tmp, prints the Rust
version contained inside and the URL to use in "distfiles_rustc_base"
to build this version.